requirements.txt
fastapi
uvicorn
pandas
openpyxl
pypdf2
python-multipart
requests
python-dotenv